+++ src/Discussion/DiscussionService.php
@@ -10,12 +10,14 @@ declare(strict_types=1);
 namespace Wizaplace\SDK\Discussion;
 
 use GuzzleHttp\Exception\ClientException;
+use GuzzleHttp\Exception\GuzzleException;
 use GuzzleHttp\RequestOptions;
 use Psr\Http\Message\UploadedFileInterface;
 use Symfony\Component\HttpFoundation\Response;
 use Wizaplace\SDK\AbstractService;
 use Wizaplace\SDK\Authentication\AuthenticationRequired;
 use Wizaplace\SDK\Catalog\DeclinationId;
+use Wizaplace\SDK\Exception\JsonDecodingError;
 use Wizaplace\SDK\Exception\NotFound;
 use Wizaplace\SDK\Exception\ProductNotFound;
 use Wizaplace\SDK\Exception\SomeParametersAreInvalid;
@@ -53,8 +55,8 @@ class DiscussionService extends AbstractService
      *
      * @return Discussion[]
      * @throws AuthenticationRequired
-     * @throws \GuzzleHttp\Exception\GuzzleException
-     * @throws \Wizaplace\SDK\Exception\JsonDecodingError
+     * @throws GuzzleException
+     * @throws JsonDecodingError
      */
     public function getDiscussions(): array
     {
@@ -77,8 +79,8 @@ class DiscussionService extends AbstractService
      *
      * @return Discussion
      * @throws AuthenticationRequired
-     * @throws \GuzzleHttp\Exception\GuzzleException
-     * @throws \Wizaplace\SDK\Exception\JsonDecodingError
+     * @throws GuzzleException
+     * @throws JsonDecodingError
      */
     public function getDiscussion(int $discussionId): Discussion
     {
@@ -94,8 +96,8 @@ class DiscussionService extends AbstractService
      *
      * @return Discussion
      * @throws AuthenticationRequired
-     * @throws \GuzzleHttp\Exception\GuzzleException
-     * @throws \Wizaplace\SDK\Exception\JsonDecodingError
+     * @throws GuzzleException
+     * @throws JsonDecodingError
      */
     public function startDiscussion(int $productId): Discussion
     {
@@ -113,8 +115,8 @@ class DiscussionService extends AbstractService
      *
      * @return Discussion
      * @throws AuthenticationRequired
-     * @throws \GuzzleHttp\Exception\GuzzleException
-     * @throws \Wizaplace\SDK\Exception\JsonDecodingError
+     * @throws GuzzleException
+     * @throws JsonDecodingError
      */
     public function startDiscussionWithVendor(int $companyId): Discussion
     {
@@ -133,8 +135,8 @@ class DiscussionService extends AbstractService
      * @return Discussion
      * @throws AuthenticationRequired
      * @throws ProductNotFound
-     * @throws \GuzzleHttp\Exception\GuzzleException
-     * @throws \Wizaplace\SDK\Exception\JsonDecodingError
+     * @throws GuzzleException
+     * @throws JsonDecodingError
      */
     public function startDiscussionFromDeclinationId(DeclinationId $declinationId): Discussion
     {
@@ -154,17 +156,24 @@ class DiscussionService extends AbstractService
      * Get the discussion's messages list
      *
      * @param int $discussionId
+     * @param null|bool $markMessagesAsRead
      *
      * @return Message[]
      * @throws AuthenticationRequired
-     * @throws \GuzzleHttp\Exception\GuzzleException
-     * @throws \Wizaplace\SDK\Exception\JsonDecodingError
+     * @throws GuzzleException
+     * @throws JsonDecodingError
      */
-    public function getMessages(int $discussionId): array
+    public function getMessages(int $discussionId, ?bool $markMessagesAsRead = null): array
     {
         $this->client->mustBeAuthenticated();
 
-        $messages = $this->client->get('discussions/' . $discussionId . '/messages');
+        $endpoint = 'discussions/' . $discussionId . '/messages';
+        if (\is_bool($markMessagesAsRead) === true) {
+            $markMessagesAsRead = $markMessagesAsRead === true ? 'true' : 'false';
+            $endpoint .= '/?markMessagesAsRead=' . $markMessagesAsRead;
+        }
+
+        $messages = $this->client->get($endpoint);
         $userId = $this->client->getApiKey()->getId();
 
         return array_map(
@@ -187,8 +196,8 @@ class DiscussionService extends AbstractService
      * @return Message
      * @throws AuthenticationRequired
      * @throws SomeParametersAreInvalid
-     * @throws \GuzzleHttp\Exception\GuzzleException
-     * @throws \Wizaplace\SDK\Exception\JsonDecodingError
+     * @throws GuzzleException
+     * @throws JsonDecodingError
      */
     public function postMessage(int $discussionId, string $content, array $files = null): Message
     {
@@ -235,8 +244,8 @@ class DiscussionService extends AbstractService
      *
      * @return DiscussionService
      *
-     * @throws \GuzzleHttp\Exception\GuzzleException
-     * @throws \Wizaplace\SDK\Exception\JsonDecodingError
+     * @throws GuzzleException
+     * @throws JsonDecodingError
      */
     public function submitContactRequest(
         string $senderEmail,
@@ -313,8 +322,8 @@ class DiscussionService extends AbstractService
      * @throws NotFound
      * @throws SomeParametersAreInvalid
      * @throws UnauthorizedModerationAction
-     * @throws \GuzzleHttp\Exception\GuzzleException
-     * @throws \Wizaplace\SDK\Exception\JsonDecodingError
+     * @throws GuzzleException
+     * @throws JsonDecodingError
      */
     public function startDiscussionWithCustomer(int $userId): Discussion
     {
@@ -332,17 +341,95 @@ class DiscussionService extends AbstractService
 
             return new Discussion($discussionData);
         } catch (ClientException $e) {
-            switch ($e->getResponse()->getStatusCode()) {
-                case Response::HTTP_BAD_REQUEST:
-                    throw new SomeParametersAreInvalid($e->getMessage());
-                case Response::HTTP_NOT_FOUND:
-                    throw new NotFound($e->getMessage());
-                case Response::HTTP_UNAUTHORIZED:
-                    throw new UnauthorizedModerationAction($e->getMessage());
+            $this->clientException($e);
+        }
+    }
 
-                default:
-                    throw $e;
-            }
+
+    /**
+     * Vendor start a discussion with a Customer from an order.
+     *
+     * @param int $orderId
+     * @param int $userId
+     *
+     * @return Discussion
+     *
+     * @throws AuthenticationRequired
+     * @throws GuzzleException
+     * @throws JsonDecodingError
+     * @throws NotFound
+     * @throws SomeParametersAreInvalid
+     * @throws UnauthorizedModerationAction
+     */
+    public function startDiscussionOnOrderWithCustomer(int $orderId, int $userId): Discussion
+    {
+        $this->client->mustBeAuthenticated();
+
+        try {
+            $discussionData = $this->client->post(
+                'discussions',
+                [
+                    RequestOptions::JSON => [
+                        'userId' => $userId,
+                        'orderId' => $orderId
+                    ]
+                ]
+            );
+
+            return new Discussion($discussionData);
+        } catch (ClientException $e) {
+            $this->clientException($e);
+        }
+    }
+
+    /**
+     * Customer start a discussion with a Customer from an order.
+     *
+     * @param int $orderId
+     * @param int $companyId
+     *
+     * @return Discussion
+     *
+     * @throws AuthenticationRequired
+     * @throws GuzzleException
+     * @throws JsonDecodingError
+     * @throws NotFound
+     * @throws SomeParametersAreInvalid
+     * @throws UnauthorizedModerationAction
+     */
+    public function startDiscussionOnOrderWithCompany(int $orderId, int $companyId): Discussion
+    {
+        $this->client->mustBeAuthenticated();
+
+        try {
+            $discussionData = $this->client->post(
+                'discussions',
+                [
+                    RequestOptions::JSON => [
+                        'companyId' => $companyId,
+                        'orderId' => $orderId
+                    ]
+                ]
+            );
+
+            return new Discussion($discussionData);
+        } catch (ClientException $e) {
+            $this->clientException($e);
+        }
+    }
+
+    private function clientException(ClientException $e): void
+    {
+        switch ($e->getResponse()->getStatusCode()) {
+            case Response::HTTP_BAD_REQUEST:
+                throw new SomeParametersAreInvalid($e->getMessage());
+            case Response::HTTP_NOT_FOUND:
+                throw new NotFound($e->getMessage());
+            case Response::HTTP_UNAUTHORIZED:
+                throw new UnauthorizedModerationAction($e->getMessage());
+
+            default:
+                throw $e;
         }
     }
 }