+++ tests/ApiClientTest.php
@@ -62,7 +62,8 @@ final class ApiClientTest extends ApiTestCase
 
         $expectedRequestOptions = [
             'headers' => [
-                'User-Agent' => 'Wizaplace-PHP-SDK/' . $version . ' PHP/' . PHP_VERSION
+                'User-Agent' => 'Wizaplace-PHP-SDK/' . $version . ' PHP/' . PHP_VERSION,
+                'Accept' => 'application/json'
             ],
         ];
 
@@ -143,7 +144,8 @@ final class ApiClientTest extends ApiTestCase
         $expectedRequestOptions = [
             'headers' => [
                 'Foo' => 'Bar',
-                'User-Agent' => 'Wizaplace-PHP-SDK/' . $version . ' PHP/' . PHP_VERSION
+                'User-Agent' => 'Wizaplace-PHP-SDK/' . $version . ' PHP/' . PHP_VERSION,
+                'Accept' => 'application/json'
             ],
         ];
 
@@ -168,6 +170,115 @@ final class ApiClientTest extends ApiTestCase
         $dispatcherMock->expects(static::once())->method('dispatchRequestEnd')->with($uniqueEventId, $response);
 
         $apiClient = new ApiClient($guzzleMock, null, $dispatcherMock);
+        self::assertSame($response, $apiClient->rawRequest('GET', $uri, $expectedRequestOptions));
+    }
+
+    public function testContentTypeOverride(): void
+    {
+        $version = PrettyVersions::getVersion('wizaplace/sdk')->getPrettyVersion();
+        $this->assertNotEmpty($version);
+
+        $expectedRequestOptions = [
+            'headers' => [
+                'Foo' => 'Bar',
+                'User-Agent' => 'Wizaplace-PHP-SDK/' . $version . ' PHP/' . PHP_VERSION,
+                'Accept' => 'customType'
+            ],
+        ];
+
+        $response = new Response();
+        $guzzleMock = $this->createMock(Client::class);
+        $uri = 'test-uri';
+
+        $guzzleMock
+            ->expects($this->exactly(1))
+            ->method('request')
+            ->withConsecutive(
+                ['GET', $uri, $expectedRequestOptions]
+            )
+            ->willReturn($response)
+        ;
+
+        $apiClient = new ApiClient($guzzleMock);
+        self::assertSame($response, $apiClient->rawRequest('GET', $uri, $expectedRequestOptions));
+    }
+
+    public function testDefaultAcceptLanguageAndContentLanguageHeadersWhenLanguageIsSet(): void
+    {
+        $version = PrettyVersions::getVersion('wizaplace/sdk')->getPrettyVersion();
+        $this->assertNotEmpty($version);
+
+        $requestHeaders = [
+            'headers' => [
+                'Foo' => 'Bar',
+                'User-Agent' => 'Wizaplace-PHP-SDK/' . $version . ' PHP/' . PHP_VERSION,
+            ],
+        ];
+
+        $expectedRequestOptions = [
+            'headers' => [
+                'Foo' => 'Bar',
+                'User-Agent' => 'Wizaplace-PHP-SDK/' . $version . ' PHP/' . PHP_VERSION,
+                'Accept-Language' => 'fr',
+                'Content-Language' => 'fr',
+                'Accept' => 'application/json'
+            ]
+        ];
+
+        $response = new Response();
+        $guzzleMock = $this->createMock(Client::class);
+        $uri = 'test-uri';
+
+        $guzzleMock
+            ->expects($this->exactly(1))
+            ->method('request')
+            ->withConsecutive(
+                ['GET', $uri, $expectedRequestOptions]
+            )
+            ->willReturn($response)
+        ;
+
+        $apiClient = new ApiClient($guzzleMock);
+        $apiClient->setLanguage('fr');
+
+        self::assertSame($response, $apiClient->rawRequest('GET', $uri, $requestHeaders));
+    }
+
+    public function testCustomAcceptLanguageAndContentLanguageHeaders(): void
+    {
+        $version = PrettyVersions::getVersion('wizaplace/sdk')->getPrettyVersion();
+        $this->assertNotEmpty($version);
+
+        $expectedRequestOptions = [
+            'headers' => [
+                'Foo' => 'Bar',
+                'User-Agent' => 'Wizaplace-PHP-SDK/' . $version . ' PHP/' . PHP_VERSION,
+                'Accept-Language' => 'es',
+                'Content-Language' => 'it',
+                'Accept' => 'application/json'
+            ],
+        ];
+
+        $response = new Response();
+        $guzzleMock = $this->createMock(Client::class);
+        $uri = 'test-uri';
+
+        $guzzleMock
+            ->expects($this->exactly(2))
+            ->method('request')
+            ->withConsecutive(
+                ['GET', $uri, $expectedRequestOptions]
+            )
+            ->willReturn($response)
+        ;
+
+        $apiClient = new ApiClient($guzzleMock);
+
+        // First, without any defined language
+        self::assertSame($response, $apiClient->rawRequest('GET', $uri, $expectedRequestOptions));
+
+        // Next, with a defined language
+        $apiClient->setLanguage('de');
         self::assertSame($response, $apiClient->rawRequest('GET', $uri, $expectedRequestOptions));
     }
 }