+++ CHANGELOG.md
@@ -1,6 +1,27 @@
 
 # Changelog
 
+## 1.147.0
+
+### New features
+
+- Set Content-Language headers on API requests
+- Updated `\Wizaplace\SDK\Discussion\DiscussionService::getMessages` add 1 param `$markMessagesAsRead`
+
+### Bugfix
+- Removed authentication required when asking to get all the available payment methods for a given basket.
+
+## 1.146.0
+
+### New features
+
+- Added optional parameter `$companyIds` to `Wizaplace\SDK\Pim\Product\ProductService::updateStock`
+- In the `Wizaplace\SDK\Discussion\Discussion` class, added the "orderId" field which is returned by the API endpoint.
+- Added `Wizaplace\SDK\Discussion\DiscussionService::startDiscussionOnOrderWithCustomer`
+- Added `Wizaplace\SDK\Discussion\DiscussionService::startDiscussionOnOrderWithCompany`
+- Fixed some @param comments which were plurals
+- Added a default _Accept_ header set to _application/json_.
+
 ## 1.144.0
 
 ### New features